Add patch to fix duplicate links due to code
Sometimes you want to demonstrate links in markdown documents by
protecting them with code blocks. This unfortunately creates a situation
where a link matches twice and you need to choose. Since these links are
not present in the XML in the first place, it's safe to choose a single
link of the two that has sunk the furthest.