remove namespace from XML.
This removes the namespace from the internal XML representation,
meaning that it will be easier to work with xml nodes because a prefix
is not needed to select nodes from the document.

This is a _breaking change_ because users of tinkr will have been using
the `md:` or `d1:` prefixes, which will no longer work.

Luckily, `md:` and `d1:` are easy to find and remove with sed in a
codebase and the `md_ns()` function can remain as is because having a
namespace definition for a document without a namespace does not prevent
anything from moving forward.