first working version of protected escape text.
In this version, we no longer need to split nodes in order to protect
them if we also want them to be continuous. I've taken the XSL template
"escape-text" and modified it so that it takes in three new parameters:

1. `pos`..........the position of the current character
2. `protect.pos`..a space-separated list of starting positions for
   protection
3. `protect.end`..a space-separated list of ending positions for protection

I've also added three new helper templates to handle list contents:

`peek` returns the top of the list, `trim` trims off the first element
of the list (or returns the value if it's not a list), and `adjust-range`
trims a list depending on if the current value is within range.

There's a lot of printing here because I wasn't too confident with
debugging, but based on my test in inst/extdata/xml_protect.xml, it
produces results correctly.