add note for creating a test
I have been able to get the error to appear with the original math
example by removing one of the dollar signs, but I was unable to
reproduce it using just the header text. I do not have time to debug
this fully right now, so I added a note to myself to come back to this.